fix bug

+++ ruoyi/src/main/java/com/ruoyi/project/performance/service/impl/PerforNormalServiceImpl.java
@@ -1,9 +1,9 @@
 package com.ruoyi.project.performance.service.impl;
 
+import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import com.ruoyi.project.performance.domain.NormalTask;
 import com.ruoyi.project.performance.domain.PerforNormal;
@@ -22,21 +22,26 @@ public class PerforNormalServiceImpl implements IPerforNormalService {
   @Override
   @Transactional
   public int insertNormalTasks(PerforNormal normal) {
-    Date date = normal.getDate();
-    String type = normal.getType();
-    String createBy = normal.getCreateBy();
-    perforNormalMapper.deleteNormalTasks(type,date);
-    List<NormalTask> normalTasks = normal.getEvaluations().stream().filter(task -> task.getOwnerId() != null).map(task -> {
-      task.setDate(date);
-      task.setType(type);
-      task.setCreateBy(createBy);
-      return task;
-    }).collect(Collectors.toList());
-    return perforNormalMapper.insertNormalTasks(normalTasks);
+    if(normal.getEvaluations().size()>0){
+      Date date = normal.getDate();
+      String type = normal.getType();
+      String createBy = normal.getCreateBy();
+      perforNormalMapper.deleteNormalTasks(type, date);
+      List<NormalTask> normalTasks = normal.getEvaluations().stream().filter(task -> task.getOwnerId() != null).map(task -> {
+        task.setDate(date);
+        task.setType(type);
+        task.setCreateBy(createBy);
+        return task;
+      }).collect(Collectors.toList());
+      return perforNormalMapper.insertNormalTasks(normalTasks);
+    }else {
+      return 1;
+    }
+
   }
 
   @Override
   public List<NormalTask> selectPerforNormalTasks(String type, Date date) {
-    return perforNormalMapper.selectPerforNormalTasks(type,date);
+    return perforNormalMapper.selectPerforNormalTasks(type, date);
   }
 }
